requirements.txt
apscheduler
cloudinary
supabase
python-dotenv
shot-scraper